Clean up suites
Switch to tasty, split up benchmark/test suite
