+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "uap-core"]
-	path = uap-core
+	path = deps/uap-core
+	url = git@github.com:ua-parser/uap-core.git
+[submodule "deps/uap-core"]
+	path = deps/uap-core
 	url = git@github.com:ua-parser/uap-core.git
